Changed unnecessary int -> Long -> int conversion (#1637)

+++ src/main/java/redis/clients/jedis/BinaryClient.java
@@ -102,7 +102,7 @@ public class BinaryClient extends Connection {
         getStatusCodeReply();
       }
       if (db > 0) {
-        select(Long.valueOf(db).intValue());
+        select(db);
         getStatusCodeReply();
       }
     }
